fix: Windows unit tests that exercised color
Again, WIN and NIX behave diff. On NIX, the ansi codes are part of the msg and the terminal transforms them into color. Therefore, we can assert that the ansi codes are in the msg. On WIN, Jansi intercepts those ansi codes and transforms them into color. So when it reaches the System.out there are no ansi codes, hence we can't assert their presence in the message. If Jansi is broken, THEN we will see ansi codes in the msg. That's why I inverted the assertions for WIN tests.